+++ CHANGES.rst
@@ -4,6 +4,7 @@
 Added
 ^^^^^
 - OIDC ``amr`` claim support. :issue:`302`
+- Server-side sessions.
 
 [0.1.0] - 2025-11-13
 --------------------